Move test class into common part and add deviceType
